requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fonttools==4.56.0
h5py==3.13.0
idna==3.10
imageio==2.25.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.1
nptyping==2.5.0
numba==0.61.0
numpy==1.26.4
packaging==24.2
pillow==10.3.0
platformdirs==4.3.7
pooch==1.8.2
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyvista==0.44.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
scooby==0.10.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
vtk==9.3.1
y-tal==0.17.2
